refactor: gameProcess.js 코드 startGame.js, utils.js로 이동
