+++ js/controller/startGame.js
@@ -1,13 +1,14 @@
 import { store } from "../model.js";
 import { renderWinner } from "./renderWinner.js";
-import {
-  gameProcess,
-  paintSpinnerIcon,
-  deleteSpinnerIcon,
-} from "./gameProcess.js";
 import { winnerMessage } from "../view/winnerTemplate.js";
-import { $ } from "../constants/DOM.js";
-import { gameProcessTemplate } from "../view/gameProcessTemplate.js";
+import { $, $$ } from "../constants/DOM.js";
+import {
+  forwardIconTemplate,
+  gameProcessTemplate,
+  spinnerIconTemplate,
+} from "../view/gameProcessTemplate.js";
+import { getRandomNumber } from "./utils.js";
+import { GAME } from "../constants/constants.js";
 
 export const startGame = () => {
   carNameRender();
@@ -25,10 +26,39 @@ export const startGame = () => {
   }, 1000);
 };
 
-export const carNameRender = () => {
+const carNameRender = () => {
   const app = $("#app");
   const div = document.createElement("div");
 
   div.innerHTML = gameProcessTemplate(store.players);
   app.appendChild(div);
+};
+
+const gameProcess = () => {
+  const cars = store.players;
+  cars.forEach((car) => {
+    if (isEffectiveScore(getRandomNumber(GAME.MIN_SCORE, GAME.MAX_SCORE))) {
+      car.point++;
+      const $carWrap = $(`[data-id="${car.name}"]`);
+      $carWrap.insertAdjacentHTML("afterend", forwardIconTemplate);
+    }
+  });
+};
+
+const isEffectiveScore = (num) => {
+  return num >= GAME.EFFECTIVE_SCORE;
+};
+
+const paintSpinnerIcon = () => {
+  const players = $$(".car-player");
+  players.forEach((player) => {
+    player.insertAdjacentHTML("afterend", spinnerIconTemplate);
+  });
+};
+
+const deleteSpinnerIcon = () => {
+  const spinners = $$("#spinnerIcon");
+  spinners.forEach((spinner) => {
+    spinner.remove();
+  });
 };
